+++ sketch-core/src/main/kotlin/com/github/panpf/sketch/resize/LongImageDecider.kt
@@ -17,7 +17,9 @@ package com.github.panpf.sketch.resize
 
 import com.github.panpf.sketch.util.format
 
-// todo add LongImageDecider()
+fun LongImageDecider(): LongImageDecider {
+    return DefaultLongImageDecider()
+}
 
 /**
  * Determine whether it is a long image given the image size and target size